test(connect-popup): move log-page and popup-error-page to new popup-pages
